Refactored API server routes and file structure
