+++ server/app.js
@@ -5,16 +5,11 @@ var logger = require('morgan');
 var cookieParser = require('cookie-parser');
 var bodyParser = require('body-parser');
 
-var index = require('./routes/index');
-var companies = require('./routes/companies/index');
-var events = require('./routes/events/index');
-var users = require('./routes/users/index');
-
 var app = express();
 
-// // view engine setup
-// app.set('views', path.join(__dirname, 'views'));
-// app.set('view engine', 'jade');
+// Register API routes in app
+var routes = require('./routes');
+routes(app)
 
 // uncomment after placing your favicon in /public
 //app.use(favicon(path.join(__dirname, 'public', 'favicon.ico')));
@@ -26,27 +21,20 @@ app.use(cookieParser());
 // Static path to production build of React Client
 app.use(express.static(path.join(__dirname, '../client/build')));
 
-app.use('/', index);
-app.use('/companies', companies);
-app.use('/events', events);
-app.use('/users', users);
-
-// catch 404 and forward to error handler
+// Catch 404 and forward to error handler
 app.use(function(req, res, next) {
   var err = new Error('Not Found');
   err.status = 404;
   next(err);
 });
 
-// error handler
+// Error handler
 app.use(function(err, req, res, next) {
-  // set locals, only providing error in development
-  res.locals.message = err.message;
-  res.locals.error = req.app.get('env') === 'development' ? err : {};
-
-  // render the error page
-  res.status(err.status || 500);
-  res.render('error');
+  const error_msg = {
+    status: err.status || 500,
+    message: err.message
+  };
+  res.json(error_msg)
 });
 
 module.exports = app;
