+++ server/routes/companies/index.js
@@ -1,3 +1,4 @@
+/* Route Prefix: /companies */
 var express = require('express');
 var router = express.Router();
