+++ server/routes/events/index.js
@@ -1,3 +1,4 @@
+/* Route Prefix: /events */
 var express = require('express');
 var router = express.Router();
