+++ server/routes/users/index.js
@@ -1,3 +1,4 @@
+/* Route Prefix: /users */
 var express = require('express');
 var router = express.Router();
